Changed internal clock to store the time in milliseconds, added SERV.TIMEHIRES, SERV.TICKPERIOD. Fixed ShipSpeed.period, made some keywords RO.
- Changed: Sphere internal clock now stores the time in milliseconds, instead of in tenth of second (which was good until the amount of ticks per second was fixed to 10). This makes the internal timer independent from the ticks per second and also should solve Issue #132.
 SERV.TIME is still returning the internal time in tenth of seconds, to keep backwards compatibility.
- Added: SERV.TIMEHIRES, which returns the internal timer in milliseconds.
- Added: SERV.TICKPERIOD, to show how much server ticks are in a real-world second (right now it's 4, so that the server ticks every 0.25 seconds).
- Fixed: ShipSpeed.period (tenths of second between each ship movement) not working as expected.
- Fixed: While sailing on the debug build, flashing spots were supposed to be shown where the next movement would move the ship (so, if moving on those tiles would make a valid movement).
- Changed: Made Read-Only SERV.CREATE, SERV.LASTVALIDDATE, SERV.LASTVALIDTIME, CREATE char and item property.

+++ src/common/CSFileObjContainer.h
@@ -16,7 +16,7 @@ class CSFileObjContainer : public CScriptObj
 private:
     std::vector<CSFileObj *> sFileList;
     int iFilenumber;
-    int iGlobalTimeout;
+    int64 iGlobalTimeout;   // in milliseconds
     int iCurrentTick;
     // ----------- //
     static lpctstr const sm_szLoadKeys[];
